+++ CHANGELOG.md
@@ -6,6 +6,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2025-05-19
+
+### Added
+
+- TextSnippets:
+  - Added a new post type for the TextSnippets feature
+  - Added a new shortcode for the TextSnippets feature
+  - Added a new bricks dynamic tag for the TextSnippets feature
+
+- ContactInfos:
+  - Added a new bricks dynamic tag for the ContactInfos feature
+
 ## [0.3.2] - 2025-05-16
 
 ### Changed